feat(LoginViewModel): Added navigation to home after authentication
