+++ app/src/main/java/com/example/dogapp/data/AppointmentsDB.kt
@@ -1,11 +1,11 @@
-package data
+package com.example.dogapp.data
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import model.Appointment
-import utils.Constants.NAME_DB
+import com.example.dogapp.model.Appointment
+import com.example.dogapp.utils.Constants.NAME_DB
 
 @Database(entities = [Appointment::class], version = 1)
 abstract class AppointmentsDB: RoomDatabase() {
