+++ app/src/main/java/com/example/dogapp/model/Appointment.kt
@@ -1,4 +1,4 @@
-package model
+package com.example.dogapp.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
