+++ app/src/main/java/com/example/dogapp/model/Symptom.kt
@@ -1,6 +1,6 @@
-package model
+package com.example.dogapp.model
 
-data class Breed(
+data class Symptom (
     val name: String
 ) {
     override fun toString(): String {
