+++ app/src/main/java/com/example/dogapp/view/adapter/breedsAdapter.kt
@@ -6,15 +6,14 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.TextView
-import androidx.navigation.NavController
 import com.example.dogapp.R
-import model.Breed
+import com.example.dogapp.model.Breed
 
 class breedsAdapter(
     context: Context,
     breeds: MutableList<Breed>,
 
-) :ArrayAdapter<Breed>(context, 0, breeds) {
+    ) :ArrayAdapter<Breed>(context, 0, breeds) {
     override fun getView(position: Int, convertView: View?, parent: ViewGroup): View {
         val view = convertView ?: LayoutInflater.from(context).inflate(R.layout.item_breed, parent, false)
         val breed = getItem(position)
