+++ app/src/main/java/com/example/dogapp/webservice/ApiUtils.kt
@@ -1,8 +1,8 @@
-package webservice
+package com.example.dogapp.webservice
 
 class ApiUtils {
     companion object{
-        fun getApiService(): ApiService{
+        fun getApiService(): ApiService {
             return RetrofitClient.getRetrofit().create(ApiService::class.java)
         }
     }
