+++ app/src/main/java/com/example/dogapp/webservice/RetrofitClient.kt
@@ -1,8 +1,8 @@
-package webservice
+package com.example.dogapp.webservice
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import utils.Constants.BASE_URL
+import com.example.dogapp.utils.Constants.BASE_URL
 
 object RetrofitClient {
     fun getRetrofit(): Retrofit {
